Here are two further changes:
(1) The handling of holes in our triangulation method EarClippingWithHoles is improved.
(2) It is a first step of handling holes in the Cut_Mesh part of our Cut-library.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@17962 060e22d8-9d1c-0410-868a-e3a950d63d95

+++ src/drt_cut/cut_side.cpp
@@ -1084,7 +1084,7 @@ bool GEO::CUT::Side::HoleOfFacet( Facet & facet, const std::vector<Cycle> & hole
   LINALG::Matrix<3,1> holepointlocalcoord;
   hole[0]()[0]->Coordinates( holepointcoord.A() );
   LocalCoordinates( holepointcoord, holepointlocalcoord, false );
-  int epsilon = 0;
+  double epsilon = 0;
   while ( intersectioninpoint )
   {
     intersectioninpoint = false;
